allin1tool/requirements.txt
pytesseract==0.3.8
Flask-SQLAlchemy==2.4.1
mysqlclient==2.1.0
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
Pillow==8.4.0
python-dateutil==2.8.2
Cython==0.29.24
click==7.1.2
MarkupSafe==2.0
Werkzeug==2.0
psycopg2==2.9.3




